refactor contact form
