+++ src/hooks/emailNotification/useError.js
@@ -1,0 +1,5 @@
+import { useSelector } from 'react-redux';
+
+export function useError() {
+  return useSelector((state) => state.emailNotification.error);
+}
