+++ src/hooks/emailNotification/useLoaded.js
@@ -1,0 +1,5 @@
+import { useSelector } from 'react-redux';
+
+export function useLoaded() {
+  return useSelector((state) => state.emailNotification.loaded);
+}
