+++ src/hooks/emailNotification/useLoading.js
@@ -1,0 +1,5 @@
+import { useSelector } from 'react-redux';
+
+export function useLoading() {
+  return useSelector((state) => state.emailNotification.loading);
+}
